Optimized the calculation of moves and their receipt by the renderer.
Created the class "Vector". The vector stores information about the piece to which it belongs, contains information about the cells available or blocked for the move. The result of calculating the piece's moves is stored in a vector.

The vectors of all pieces are stored in an object of the "Moves" class. This class stores methods for checking castling, checkmate, check, stalemate, mate threat. The class also contains a method for updating information about moves after the players' move, and only moves that have become invalid are recalculated.

And ... and it also returns the moves of a specific chess piece or all the moves of a player of a certain color.